refactor(structure): remove necessity to export from main.ts
Now all source files are bundled automatically. This will keep folder structures of /src and /test more in sync
